fix: 🐛 time change and fixed nit
